Moved to a single shader to avoid code duplication
