+++ packages/text-bitmap/src/index.ts
@@ -3,4 +3,3 @@ export * from './BitmapFontLoader';
 export * from './BitmapFont';
 export * from './BitmapFontData';
 export * from './BitmapTextStyle';
-export * from './MSDFText';
